+++ serial_config.h
@@ -161,7 +161,7 @@ private:
     static int GetInt(const Json::Value& obj, const std::string& key);
     static int ToInt(const Json::Value& v, const std::string& title);
     static uint64_t ToUint64(const Json::Value& v, const std::string& title);
-    static std::tuple<int, uint8_t, uint8_t> ParseRegisterAddress(const Json::Value& obj, const std::string& key);
+    static std::tuple<int, uint16_t, uint8_t> ParseRegisterAddress(const Json::Value& obj, const std::string& key);
 
     std::string ConfigFileName;
     PHandlerConfig HandlerConfig;
